requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-cors-headers==4.3.1
django-filter==25.2
djangorestframework==3.14.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
Pillow==10.4.0
pytz==2025.2
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
